backend/requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
